data picker table separation
