+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "WestChart_ReactApp",
-  "version": "0.1.2",
+  "version": "0.1.3",
   "private": true,
   "devDependencies": {
     "autoprefixer": "6.5.1",
@@ -55,9 +55,11 @@
   },
   "dependencies": {
     "express": "latest",
+    "moment": "^2.17.1",
     "morgan": "^1.7.0",
     "react": "^15.4.2",
     "react-bootstrap": "^0.30.7",
+    "react-datepicker": "^0.40.0",
     "react-dom": "^15.4.2",
     "react-table": "^4.2.7",
     "request": "latest"
